// src/model/trajet.ts
export interface Trajet {
  departement: number;
  travail_longitude: number;
  '2009_inter': number;
  travail_departement: number;
  '2014_extra': number;
  distance_oiseau_km: number;
  duree_auto_minutes: number;
  unite_conso_menages: number;
  latitude: number;
  '2014': number;
  travail_emplois: number;
  '2009_extra': number;
  commune: string;
  emplois: number;
  revenu_median: number;
  '2009_extra_travail_commune': number;
  duree_velo_minutes: number;
  '2014_extra_travail_commune': number;
  '2014_inter': number;
  pers_par_menages: number;
  insee: number;
  distance_velo_km: number;
  distance_auto_km: number;
  travail_commune: string;
  longitude: number;
  travail_insee: number;
  travail_latitude: number;
  menages: number;
  '2009': number;
  habitants: number;
}